Fixed Loop not catching desper.Switch exceptions
In particular, if the Switch exception were dispatched from a pyglet event, it could not be catched. This is due to the fact that pyglet events propagate directly out of pyglet.app.run(). For this reason, the main loop now wraps the entire pyglet.app.run() in the Switch try. Careful reset of the state then is necessary to prevent terrible disasters (like window redraw scheduled multiple times, etc.). Luckily, the current solution seems to be stable.