Add PositionSync2D and PositionRotationSync2D
Sync position and rotation for graphical components. Especially useful for shapes, which do not support scale in general (each shape has its own way of managing "scale", like the height/width of a rectangle or the radius of a circle), and sometimes do not support rotation either.

Specialized functions that initialize the proper sync component for each shape based on its supported transformations will be defined.