Add retrieve_batch(...), obtain Batch from a World
Since most graphical components in a given World will want to be in the same batch, retrieving a "default" batch from the current (or specific) world  comfortably is desirable. This function queries for a pyglet-desper.Camera or a loose Batch component in the given world (current by default), and creates one (an entity consisting of the Batch only) if not found.